refactor: 크롤링 로직 트랜잭션 외부로 분리 및 self-injection 제거
