+++ src/main/java/com/idear/backend/contest/crawler/service/ContestSaveService.java
@@ -2,14 +2,11 @@ package com.idear.backend.contest.crawler.service;
 
 import com.idear.backend.contest.crawler.parser.ContestDetailParser;
 import com.idear.backend.contest.domain.Contest;
-import com.idear.backend.contest.repository.ContestRepository;
 import com.idear.backend.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
@@ -19,7 +16,7 @@ import java.util.Set;
 @Service
 public class ContestSaveService {
 
-  private final ContestRepository contestRepository;
+  private final ContestPersistenceService contestPersistenceService;
   private final ContestDetailParser detailParser;
 
   private static final int BATCH_SIZE = 50;
@@ -27,7 +24,6 @@ public class ContestSaveService {
   /**
    * 공모전 저장 (중복 체크 포함)
    */
-  @Transactional
   public boolean saveContestIfNotExists(String linkareerUrl, Set<String> processedUrls) {
     // 메모리 중복 체크
     if (processedUrls.contains(linkareerUrl)) {
@@ -39,18 +35,8 @@ public class ContestSaveService {
       // 상세 정보 크롤링
       Contest contest = detailParser.parseDetailPage(linkareerUrl);
 
-      // DB 중복 체크
-      if (isDuplicate(contest)) {
-        processedUrls.add(linkareerUrl);
-        return false;
-      }
-
       // 저장
-      contestRepository.save(contest);
-      processedUrls.add(linkareerUrl);
-      log.info("공모전 저장: {}", contest.getTitle());
-
-      return true;
+      return contestPersistenceService.saveIfNotDuplicate(contest, linkareerUrl, processedUrls);
 
     } catch (CustomException e) {
       log.error("크롤링 실패: {}", linkareerUrl, e);
@@ -62,7 +48,6 @@ public class ContestSaveService {
   /**
    * 공모전 배치 저장 (여러 건을 한 번에)
    */
-  @Transactional
   public int saveBatch(List<String> urls, Set<String> processedUrls) {
     List<Contest> contestBatch = new ArrayList<>();
     int savedCount = 0;
@@ -78,21 +63,16 @@ public class ContestSaveService {
         // 상세 정보 크롤링
         Contest contest = detailParser.parseDetailPage(url);
 
-        // DB 중복 체크
-        if (!isDuplicate(contest)) {
+        if (!contestPersistenceService.isDuplicate(contest)) {
           contestBatch.add(contest);
-
-          // 배치 크기에 도달하면 저장
-          if (contestBatch.size() >= BATCH_SIZE) {
-            contestRepository.saveAll(contestBatch);
-            savedCount += contestBatch.size();
-            log.info("배치 저장 완료: {}건", contestBatch.size());
-            contestBatch.clear();
-          }
         }
 
         processedUrls.add(url);
 
+        if (contestBatch.size() >= BATCH_SIZE) {
+          savedCount += contestPersistenceService.saveBatchWithRetry(contestBatch);
+          contestBatch.clear();
+        }
       } catch (CustomException e) {
         log.error("크롤링 실패: {}", url, e);
         processedUrls.add(url);
@@ -101,70 +81,9 @@ public class ContestSaveService {
 
     // 남은 데이터 저장
     if (!contestBatch.isEmpty()) {
-      savedCount += saveBatchWithRetry(contestBatch);
+      savedCount += contestPersistenceService.saveBatchWithRetry(contestBatch);
     }
 
     return savedCount;
-  }
-
-  private int saveBatchWithRetry(List<Contest> contestBatch) {
-    try {
-      contestRepository.saveAll(contestBatch);
-      log.info("배치 저장 완료: {}건", contestBatch.size());
-      return contestBatch.size();
-
-    } catch (Exception e) {
-      log.error("배치 저장 실패, 개별 저장으로 폴백: {}건", contestBatch.size(), e);
-      return saveIndividually(contestBatch);
-    }
-  }
-
-  /**
-   * 개별 저장 (폴백 전략)
-   */
-  private int saveIndividually(List<Contest> contests) {
-    int saved = 0;
-
-    for (Contest contest : contests) {
-      try {
-        contestRepository.save(contest);
-        saved++;
-        log.debug("개별 저장 성공: {}", contest.getTitle());
-      } catch (Exception e) {
-        log.error("개별 저장 실패: {} - {}", contest.getTitle(), e.getMessage());
-      }
-    }
-
-    log.info("개별 저장 폴백 완료: {}개 중 {}개 성공", contests.size(), saved);
-    return saved;
-  }
-
-  /**
-   * 마감된 공모전 삭제
-   */
-  @Transactional
-  public int deleteClosedContests() {
-    LocalDate today = LocalDate.now();
-    int deleted = contestRepository.deleteClosedContests(today);
-
-    if (deleted > 0) {
-      log.info("마감된 공모전 {}개 삭제", deleted);
-    }
-
-    return deleted;
-  }
-
-  /**
-   * 중복 체크 (homepageUrl 기준)
-   */
-  private boolean isDuplicate(Contest contest) {
-    String homepageUrl = contest.getHomepageUrl();
-
-    if (homepageUrl != null && contestRepository.existsByHomepageUrl(homepageUrl)) {
-      log.debug("중복된 공모전 (homepageUrl): {}", homepageUrl);
-      return true;
-    }
-
-    return false;
   }
 }
